Complete q_reverse and fix segmentation fault
The result of `make test` at this commit:
---     TOTAL           100/100

In q_insert_* functions, I free the newh and return false if I cannot
allocate space to newh->value to make sure there is no memory leak.

In q_remove_head, I make some change to sure the size of sp will not
out of the bufsize and append \0 to the end.